Moved image deletion to 'after' trigger
Now creates a list of files imported during the 'before' trigger,
and deletes them both from the flash card and the filesystem during
the 'after' trigger.  This means that the core Lychee import doesn't
need to delete the files, but this plugin can still ensure the same
image isn't imported multiple times.